Add a special JSX component for binding Children
Because we want to support dynamic child binding in all/most components, rendering/binding children needs to be a dynamic component itself, but also because it is a deep internal concern of the renderer itself it forms a special case in the JSX "Description" tree akin to Fragment.

To support "deep" binding of children (a component passes its children into a "render function" or equivalent to a child component), the Children component takes an optional ComponentContext.